Time: 0 ms (100.00%), Space: 6 MB (66.67%) - LeetHub

+++ 2579-count-total-number-of-colored-cells/2579-count-total-number-of-colored-cells.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     long long coloredCells(int n) {
-        long long Sum = 1;
-        long long Previous = 0;
+        long Sum = 1;
+        int Previous = 0;
         
         for (int i = 2; i <= n; i++)
         {
